Revert "Feature/fix5 haishin"
